+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 🦟 Bugs Fixed
 
 -   Add missing Silicon recipes to the Assembler and Atomic Forge [(\#276)](https://github.com/EnigmaticaModpacks/Enigmatica10/issues/276)
+-   Add missing Fluix dust recipe to the Pulverizer [(\#276)](https://github.com/EnigmaticaModpacks/Enigmatica10/issues/276)
 
 ---